Separate a manager's life into stages
A manager's life cycle is now split into obvious setup, call and cleanup
stages.  This allows us to use setup_call_cleanup/3 to guarantee that
cleanup is carried out no matter how the manager ends.

Cleanup is not much better than it was before, but now we have room to
expand it.